add support for matching with `HasPrefix` to `FilteredRelayConfig`
This change adds a new `HasPrefix` field to the `FilteredRelayConfig`,
which allows a relay endpoint to match & forward a request only if the
value at the given JSON path in the request's body starts with at least
one of the configured strings defined by the `HasPrefix` rule.

This allows users to create relay configurations which, for example,
will only forward requests with a comment body that starts with
"/semgrep", while dropping/ignoring requests with a comment body that
merely _contains_ "/semgrep" (for example if someone is quoting an
earlier comment in the review thread, or instructing on how to use the
feature).